requirements.txt
fastapi
uvicorn
bcrypt
google-genai
